worked on fixing bugs when user tries to hit a protected api route. Implimented token storage via localstorage and rewrote auth middleware.
