+++ client/src/components/HomePage/HomePage.jsx
@@ -10,7 +10,7 @@ export default function HomePage() {
   const [itemId, setItemId] = useState()
   const [quantity, setQuantity] = useState(0)
   const [foodList, setFoodList] = useState([{}])
-  const {user} = useContext(UserContext)
+
   // Grabs food information on mount
   useEffect(() => {
     const fetchFoodData = async () => {
@@ -27,7 +27,8 @@ export default function HomePage() {
 
   // Adds item to cart
   const addToCart = async () => {
-    
+    const token = localStorage.getItem('token');
+
     if (!token) {
       console.error('Token not found');
       return;
@@ -35,11 +36,10 @@ export default function HomePage() {
 
     const headers = {
       'Content-Type': 'application/json',
-      'Authorization': `Bearer ${user.token}`,
+      'Authorization': `Bearer ${token}`,
     };
-
     try {
-      const response = await axios.post('/cart', {itemId, quantity}, {
+      const response = await axios.get('/cart',  {
         withCredentials: true,
         headers: headers,
       });
