+++ server/routes/api/userRoutes.js
@@ -7,7 +7,7 @@ const auth = require("../../middleware/auth.js");
 router.post("/register", registerUser);
 router.post("/login", loginUser);
 router.post("/logout", logoutUser);
-router.get("/me", getMe);
+router.get("/me", auth, getMe);
 
 // Export the router
 module.exports = router;
